requirements.txt
gradio==5.8.0
numpy==1.26.4
pillow==10.4.0
tflite-runtime==2.14.0
huggingface_hub==1.0.1